Dhrystone: Removed unused compiler options
The original test had options, guarded by macros, for different
compilers. These were options that handled things like whehter or not
the compiler had structs or enums. These have been fairly normal
features of C compilers for some time now and are assumed to be
availible during the compilation of tests for the test framework.

Change-Id: Iadcef998dadfc30d888c2aaeacd227deab1bf5be