This adds a policy that performs absolutely no checks.
This policy makes use of the __NO_CHECKS policy-language construct added in a
recent version of the policy tool.  This commit also does a bit of
reorganization, giving the innaccurately named "none" and "nop" policies the
more descriptive names "testSimple" and "testComplex".  The "none" policy has
been replaced by the policy that actually does no checking.