Dhrystone: Add Makefiles to customize the build process for the Dhrystone Benchmark
The added Makefiles will make use of the included linkerscript, which specifies a larger
RAM region than is normally used. This requires the use of an altered emulation model with
this added memory.
Extra LDFLAGS are added in case the ISP_PREFIX environment variable
was not set properly when the toolchain whas built and installed. This
catches the situation when there wasn't a trailing slash for the
ISP_PREFIX environment variable

Change-Id: I2338c9c8f8b950bc02a4df0e59f49864740313ef